# Fixes #14428 - improves performance with purge empty errata and groups