--- a/app/lib/actions/pulp/repository/purge_empty_errata.rb
+++ b/app/lib/actions/pulp/repository/purge_empty_errata.rb
@@ -8,16 +8,10 @@ module Actions
 
         def invoke_external_task
           repo = ::Katello::Repository.where(:pulp_id => input[:pulp_id]).first
+          errata_to_delete = repo.empty_errata
 
-          package_lists = repo.package_lists_for_publish
-          filenames = package_lists[:filenames]
-
-          errata_to_delete = repo.errata.collect do |erratum|
-            erratum.errata_id if filenames.intersection(erratum.packages.pluck(:filename)).empty?
-          end
-          errata_to_delete.compact!
           repo.unassociate_by_filter(::Katello::ContentViewErratumFilter::CONTENT_TYPE,
-                                 "id" => { "$in" => errata_to_delete })
+                                 "id" => { "$in" => errata_to_delete.map(&:errata_id) })
         end
       end
     end