@jlsherrill 2016-04-13
@ehelms and @daviddavis i'd like to highlight that I disabled the ClassLength cop here.  For this class I might could find some unused methods and randomly delete them randomly or disable it for this class only.  I chose to disable it across the board as:

* No one writes a huge class, they grow organically over time.  We generally know when a class is too big and don't necessarily need rubocop to tell us.
* someone making one small change isn't going to completely refactor or re-architect a class in order to get one change in.

curious your thoughts.  There is an existing issue open IIRC for re-factoring repository into STI