@beav 2016-04-13
@jlsherrill this looks good from a code perspective, will test now.

I agree about the ClassLength cop, I've been in similar situations where I had to reorganize code to keep it barely under the limit. I think it's a good idea to have but in practice I just juggle code to make it pass, or add a TODO task in redmine to refactor later.

I'm :+1: for disabling the cop.