@ehelms 2016-04-13
I don't like across the board changes (especially buried inside another change like this). The class length is still a good reminder when we hit it to question why a class has finally hit that point -- is it overloaded? is the PR itself adding more than it should to to it? should this be an indicator to open a refactoring PR for a particular class? If we remove this across the board, we won't have a reminder to ask ourselves these questions when PRs are opened.