@jlsherrill 2016-04-13
@ehelms you didn't really lay out an alternative solution, are you proposing as an alternative solution that when this situation arises on a smallish-pr (like this one), that we simply add a ignore for this cop on that class (assuming there is no sensible and simple re-factor available)?