@ehelms 2016-04-13
@jlsherrill yes, and ask whether a re-factoring PR should be opened. The alternative is to just not care what this cop's intent is and move along (but I'd request turning it off in a specific PR).