@jlsherrill 2016-04-13
@ehelms my point is that no one is going to open a refactor PR immediately after their change, they may file an issue somewhere but unless there's a reason for it its not really going to get done.  Refactoring a class is hard, complex, and likely to break things in most cases IMHO.