@jlsherrill 2016-04-13
@beav @daviddavis :

After chatting with @ehelms in irc, a proposal is to:

If a small change causes a class length to go over the limit, add a rubocop exception for that class and open a redmine issue assigned to the 'refactor' tracker.  There is no expectation that the PR author will resolve the newly opened issue.

This would allow us to more easily track classes ripe for refactoring.

You guys okay with that?