@ehelms 2016-04-14
I'm working on loading in some large repositories to do some larger scale testing to see this in action.