@beav 2016-04-14
@ehelms it may be overkill for a one-time test, but if you are interested you can try out perfhammer https://github.com/komidore64/katello-perf/blob/master/perfhammer.sh