@jlsherrill 2016-04-14
Issue to break up repository class:  http://projects.theforeman.org/issues/12925